Fixed the config file
Oops... My bad. I had originally added it as a Symbolic link. Meaning it would just reference back to the original config.

I've now added it as a Hardlink, meaning, now the contents of the file will actually show up here. And the file looks like what it's supposed to.